requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pillow==11.3.0
pycryptodome==3.23.0
pygame==2.6.1
requests==2.32.4
urllib3==2.5.0
